Extract boolean operator helpers
